Updated app structure
