+++ package.json
@@ -14,8 +14,9 @@
   "license": "ISC",
   "dependencies": {
     "dotenv": "^16.0.3",
+    "express": "^4.18.2",
     "mongoose": "^6.8.0",
-    "socket.io": "^4.5.4"
+    "socket.io": "^2.5.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.20"
